change-timeを修正
